refactor: Update component imports to use TitleCard instead of Card
